+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changes
 Unreleased
 ----------
 
+- Merge ClaimsBuilder into Claims. `#98 <https://github.com/dajiaji/python-cwt/pull/98>`__
+- Rename KeyBuilder to COSEKey. `#97 <https://github.com/dajiaji/python-cwt/pull/97>`__
 - Add support for external AAD. `#94 <https://github.com/dajiaji/python-cwt/pull/94>`__
 - Make unwrap_key return COSEKey. `#93 <https://github.com/dajiaji/python-cwt/pull/93>`__
 - Fix default HMAC key size. `#91 <https://github.com/dajiaji/python-cwt/pull/91>`__